requirements.txt
Flask==2.0.1        
Flask-RESTful==0.3.9
flask-marshmallow==0.14.0
marshmallow==3.12.1
python-dotenv==0.17.1
Flask-JWT
Flask-SQLAlchemy
uwsgi
psycopg2

